+++ presentation/src/main/java/com/example/presentation/main/view/fragment/like/LikeFragment.kt
@@ -9,6 +9,7 @@ import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.presentation.R
 import com.example.presentation.databinding.FragmentLikeBinding
+import com.example.presentation.login.LoginActivity
 import com.example.presentation.main.customview.ConfirmDialog
 import com.example.presentation.main.view.adapter.LikeViewAdapter
 import com.example.presentation.main.view.fragment.like.vm.LikeViewModel
@@ -74,7 +75,13 @@ class LikeFragment : Fragment(R.layout.fragment_like) {
     }
 
     private fun showLoginLayout(isVisible: Boolean, binding: FragmentLikeBinding) {
-        binding.layoutNonMember.isVisible = isVisible
+        with(binding) {
+            layoutNonMember.isVisible = isVisible
+            btLogin.setOnClickListener {
+                val intent = Intent(requireContext(), LoginActivity::class.java)
+                startActivity(intent)
+            }
+        }
     }
 
     private fun initView(binding: FragmentLikeBinding) {
@@ -89,20 +96,10 @@ class LikeFragment : Fragment(R.layout.fragment_like) {
                         showLoginLayout(false, binding)
                     }
                     MemberStatus.NonMember -> {
-                        setAlertDialog()
                         showLoginLayout(true, binding)
                     }
                 }
             }
         }
-    }
-
-    private fun setAlertDialog() {
-        val dialog = ConfirmDialog(
-            text = "로그인이 필요한 서비스입니다\n로그인하시겠습니까?",
-        ) {
-            // 로그인 화면으로 이동
-        }
-        dialog.show(childFragmentManager, "ConfirmDialog")
     }
 }
